Refactor group chat page into modular widgets
