+++ lib/features/chat/group_chat/presentation/widgets/group_chat_message_list.dart
@@ -1,0 +1,52 @@
+import 'package:flutter/material.dart';
+
+import '../../data/models/group_message.dart';
+import 'group_message_tile.dart';
+
+class GroupChatMessageList extends StatelessWidget {
+  const GroupChatMessageList({
+    super.key,
+    required this.messages,
+    required this.scrollController,
+    required this.youLabel,
+    required this.repliesLabelBuilder,
+  });
+
+  final List<GroupMessage> messages;
+  final ScrollController scrollController;
+  final String youLabel;
+  final String Function(int) repliesLabelBuilder;
+
+  @override
+  Widget build(BuildContext context) {
+    final colorScheme = Theme.of(context).colorScheme;
+
+    return Container(
+      decoration: BoxDecoration(
+        color: colorScheme.surfaceContainerLowest,
+        borderRadius: const BorderRadius.only(
+          topLeft: Radius.circular(24),
+          topRight: Radius.circular(24),
+        ),
+      ),
+      child: ListView.builder(
+        controller: scrollController,
+        padding: const EdgeInsets.only(bottom: 24, top: 16),
+        itemCount: messages.length,
+        itemBuilder: (context, index) {
+          final message = messages[index];
+          final bool showAvatar = index == 0
+              ? true
+              : !messages[index - 1].isFromSameSender(messages[index]);
+
+          return GroupMessageTile(
+            message: message,
+            showAvatar: showAvatar,
+            youLabel: youLabel,
+            repliesLabelBuilder: repliesLabelBuilder,
+          );
+        },
+      ),
+    );
+  }
+}
